+++ airbyte-integrations/connectors/destination-clickhouse-v2/src/test/kotlin/io/airbyte/integrations/destination/clickhouse_v2/write/load/BinaryRowInsertBufferTest.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.destination.clickhouse_v2.write.direct
+package io.airbyte.integrations.destination.clickhouse_v2.write.load
 
 import com.clickhouse.client.api.Client
 import com.clickhouse.client.api.data_formats.RowBinaryFormatWriter
@@ -24,6 +24,7 @@ import io.airbyte.cdk.load.data.TimestampWithTimezoneValue
 import io.airbyte.cdk.load.data.TimestampWithoutTimezoneValue
 import io.airbyte.cdk.load.orchestration.db.TableName
 import io.airbyte.cdk.load.util.serializeToString
+import io.airbyte.integrations.destination.clickhouse_v2.write.load.BinaryRowInsertBuffer
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import io.mockk.junit5.MockKExtension
@@ -59,14 +60,20 @@ class BinaryRowInsertBufferTest {
     @BeforeEach
     fun setup() {
         every {
-            clickhouseClient.getTableSchema(Fixtures.TEST_NAME, Fixtures.TEST_NAMESPACE)
+            clickhouseClient.getTableSchema(
+                Fixtures.TEST_NAME,
+                Fixtures.TEST_NAMESPACE
+            )
         } returns schema
         buffer = BinaryRowInsertBuffer(tableName, clickhouseClient)
     }
 
     @Test
     fun `initializes the ch row binary writer`() {
-        verify { clickhouseClient.getTableSchema(Fixtures.TEST_NAME, Fixtures.TEST_NAMESPACE) }
+        verify { clickhouseClient.getTableSchema(
+            Fixtures.TEST_NAME,
+            Fixtures.TEST_NAMESPACE
+        ) }
 
         assertEquals(ClickHouseFormat.RowBinary, buffer.writer.format)
     }
