+++ airbyte-integrations/connectors/destination-clickhouse-v2/src/test/kotlin/io/airbyte/integrations/destination/clickhouse_v2/write/load/ClickhouseDirectLoaderFactoryTest.kt
@@ -2,14 +2,15 @@
  * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.destination.clickhouse_v2.write.direct
+package io.airbyte.integrations.destination.clickhouse_v2.write.load
 
 import com.clickhouse.client.api.Client
 import io.airbyte.cdk.load.command.DestinationStream
 import io.airbyte.cdk.load.orchestration.db.TableName
 import io.airbyte.cdk.load.orchestration.db.direct_load_table.DirectLoadTableExecutionConfig
 import io.airbyte.cdk.load.write.StreamStateStore
-import io.airbyte.integrations.destination.clickhouse_v2.write.RecordMunger
+import io.airbyte.integrations.destination.clickhouse_v2.write.load.ClickhouseDirectLoaderFactory
+import io.airbyte.integrations.destination.clickhouse_v2.write.transform.RecordMunger
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import io.mockk.junit5.MockKExtension
