+++ airbyte-integrations/connectors/destination-clickhouse-v2/src/test/kotlin/io/airbyte/integrations/destination/clickhouse_v2/write/load/ClickhouseDirectLoaderTest.kt
@@ -2,12 +2,14 @@
  * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.destination.clickhouse_v2.write.direct
+package io.airbyte.integrations.destination.clickhouse_v2.write.load
 
 import io.airbyte.cdk.load.data.StringValue
 import io.airbyte.cdk.load.message.DestinationRecordRaw
 import io.airbyte.cdk.load.write.DirectLoader
-import io.airbyte.integrations.destination.clickhouse_v2.write.RecordMunger
+import io.airbyte.integrations.destination.clickhouse_v2.write.load.BinaryRowInsertBuffer
+import io.airbyte.integrations.destination.clickhouse_v2.write.load.ClickhouseDirectLoader
+import io.airbyte.integrations.destination.clickhouse_v2.write.transform.RecordMunger
 import io.mockk.coVerify
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
